+++ src/@types/priceSpecification.ts
@@ -1,0 +1,8 @@
+export interface IPriceSpecification {
+  sku: string;
+  price: number;
+  originalPrice: number;
+  maxPrice: number;
+  percent: number;
+  discount: number;
+}
